+++ apps/desktop/desktop/src/main/java/bisq/desktop/main/content/settings/network/transport/TransportTypeController.java
@@ -31,8 +31,8 @@ import bisq.network.p2p.node.CloseReason;
 import bisq.network.p2p.node.Connection;
 import bisq.network.p2p.node.Node;
 import bisq.network.p2p.node.NodesById;
-import bisq.network.p2p.services.peergroup.PeerGroupManager;
-import bisq.network.p2p.services.peergroup.PeerGroupService;
+import bisq.network.p2p.services.peer_group.PeerGroupManager;
+import bisq.network.p2p.services.peer_group.PeerGroupService;
 import bisq.security.keys.KeyBundleService;
 import bisq.user.profile.UserProfileService;
 import lombok.Getter;
