+++ network/network/src/main/java/bisq/network/NetworkServiceConfig.java
@@ -27,10 +27,10 @@ import bisq.network.p2p.node.authorization.AuthorizationService;
 import bisq.network.p2p.node.transport.ClearNetTransportService;
 import bisq.network.p2p.node.transport.I2PTransportService;
 import bisq.network.p2p.services.data.inventory.InventoryService;
-import bisq.network.p2p.services.peergroup.PeerGroupManager;
-import bisq.network.p2p.services.peergroup.PeerGroupService;
-import bisq.network.p2p.services.peergroup.exchange.PeerExchangeStrategy;
-import bisq.network.p2p.services.peergroup.keep_alive.KeepAliveService;
+import bisq.network.p2p.services.peer_group.PeerGroupManager;
+import bisq.network.p2p.services.peer_group.PeerGroupService;
+import bisq.network.p2p.services.peer_group.exchange.PeerExchangeStrategy;
+import bisq.network.p2p.services.peer_group.keep_alive.KeepAliveService;
 import bisq.tor.TorTransportConfig;
 import com.typesafe.config.Config;
 import lombok.Getter;
