+++ network/network/src/main/java/bisq/network/p2p/node/PeerConnectionsManager.java
@@ -24,7 +24,7 @@ import bisq.network.p2p.node.authorization.AuthorizationService;
 import bisq.network.p2p.node.network_load.NetworkLoad;
 import bisq.network.p2p.node.transport.ServerSocketResult;
 import bisq.network.p2p.node.transport.TransportService;
-import bisq.network.p2p.services.peergroup.BanList;
+import bisq.network.p2p.services.peer_group.BanList;
 import lombok.extern.slf4j.Slf4j;
 
 import java.io.IOException;
