+++ network/network/src/main/java/bisq/network/p2p/node/ServerChannel.java
@@ -20,7 +20,7 @@ package bisq.network.p2p.node;
 import bisq.network.common.Address;
 import bisq.network.p2p.node.authorization.AuthorizationService;
 import bisq.network.p2p.node.network_load.NetworkLoad;
-import bisq.network.p2p.services.peergroup.BanList;
+import bisq.network.p2p.services.peer_group.BanList;
 import lombok.Setter;
 import lombok.extern.slf4j.Slf4j;
