+++ network/network/src/main/java/bisq/network/p2p/node/handshake/ConnectionHandshakeInitiator.java
@@ -28,7 +28,7 @@ import bisq.network.p2p.node.OutboundConnection;
 import bisq.network.p2p.node.authorization.AuthorizationService;
 import bisq.network.p2p.node.authorization.AuthorizationToken;
 import bisq.network.p2p.node.network_load.NetworkLoad;
-import bisq.network.p2p.services.peergroup.BanList;
+import bisq.network.p2p.services.peer_group.BanList;
 import lombok.Getter;
 import lombok.extern.slf4j.Slf4j;
