+++ network/network/src/main/java/bisq/network/p2p/node/handshake/ConnectionHandshakeResponder.java
@@ -29,7 +29,7 @@ import bisq.network.p2p.node.authorization.AuthorizationService;
 import bisq.network.p2p.node.authorization.AuthorizationToken;
 import bisq.network.p2p.node.envelope.NetworkEnvelopeSocketChannel;
 import bisq.network.p2p.node.network_load.NetworkLoad;
-import bisq.network.p2p.services.peergroup.BanList;
+import bisq.network.p2p.services.peer_group.BanList;
 import lombok.extern.slf4j.Slf4j;
 
 import java.io.IOException;
