+++ network/network/src/main/java/bisq/network/p2p/services/peer_group/Peer.java
@@ -15,7 +15,7 @@
  * along with Bisq. If not, see <http://www.gnu.org/licenses/>.
  */
 
-package bisq.network.p2p.services.peergroup;
+package bisq.network.p2p.services.peer_group;
 
 import bisq.common.proto.NetworkProto;
 import bisq.common.validation.NetworkDataValidation;
