+++ network/network/src/main/java/bisq/network/p2p/services/peer_group/PeerGroupStore.java
@@ -15,7 +15,7 @@
  * along with Bisq. If not, see <http://www.gnu.org/licenses/>.
  */
 
-package bisq.network.p2p.services.peergroup;
+package bisq.network.p2p.services.peer_group;
 
 import bisq.common.proto.ProtoResolver;
 import bisq.common.proto.UnresolvableProtobufMessageException;
